+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - CO: support debit notes
 
+### Fixed
+
+- IT: Company fiscal code can be the same as the VAT code.
+
 ## [v0.82.0] - 2024-07-29
 
 ### Added